renamed encoder_stuff files to not conflict with quantum/encoder
