+++ users/boboysdadda/encoder_stuff.h
